+++ patchrail/packet/__init__.py
@@ -1,0 +1,1 @@
+"""Approval packet read-model package."""